edit select berat

+++ app/Http/Controllers/AccAdminController.php
@@ -2,44 +2,44 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\User;
+use App\Http\Requests\UserStoreRequest;
 use Illuminate\Http\Request;
+use App\Models\User;
 
 class AccAdminController extends Controller
 {
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function index()
     {
         $this->authorize('viewAdmin', User::class);
 
-        $data = User::where('role', '=', 'Admin')->get();
-        return view('acc-admin.index', compact('data'));
+        $data = User::where('role', 'Admin')->get();
+        return view('accounts.admin.index', compact('data'));
     }
 
-    public function edit($id)
-    {
-        $this->authorize('updateAdmin', User::class);
-
-        $data = User::find($id);
-        return view('acc-admin.edit', compact('data'));
-    }
-
-    public function editData(Request $request, $id)
-    {
-        $this->authorize('updateAdmin', User::class);
-
-        $data = User::find($id);
-        $data->update($request->all());
-        return redirect()->route('adminAcc')->with('success', 'Data berhasil diubah');
-    }
-
-    public function createData()
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
     {
         $this->authorize('createAdmin', User::class);
 
-        return view('acc-admin.tambah');
+        return view('accounts.admin.create');
     }
 
-    public function createDataAcc(Request $request)
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
     {
         $this->authorize('createAdmin', User::class);
 
@@ -55,13 +55,64 @@ class AccAdminController extends Controller
         $validations['password'] = bcrypt($validations['password']);
 
         User::create($validations);
-        return redirect()->route('adminAcc')->with('success', 'Data berhasil ditambahkan');
+        return redirect()->route('admin.index')->with('success', 'Data berhasil ditambahkan');
     }
 
-    public function delete($id)
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
     {
-        $data = User::find($id);
+        $this->authorize('updateAdmin', User::class);
+
+        $data = User::findOrFail($id);
+        return view('accounts.admin.edit', compact('data'));
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        $this->authorize('updateAdmin', User::class);
+
+        $validations = $request->validate([
+            'nama' => 'required|max:255',
+            'alamat' => 'required|max:255',
+            'email' => 'required|email:dns|max:255|unique:users,email,'. $id,
+            'no_hp' => 'required|max:255',
+            'password' => 'required|max:255',
+            'role' => 'required|max:255',
+        ]);
+
+        $validations['password'] = bcrypt($validations['password']);
+        
+        $data = User::findOrFail($id);
+        
+        $data->update($validations);
+        return redirect()->route('admin.index')->with('success', 'Data berhasil diubah');
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        $this->authorize('destroyAdmin', User::class);
+
+        $data = User::findOrFail($id);
+
         $data->delete();
-        return redirect()->route('adminAcc');
+        return redirect()->route('admin.index');
     }
 }
